+++ AllToDo-iOS/AllToDo/Views/KakaoMapView.swift
@@ -318,25 +318,25 @@ struct KakaoMapView: UIViewRepresentable {
             
             // 2. Prepare Data
             var allItemsToProcess: [UnifiedMapItem] = []
-            var rawPoints: [Int32] = []
+            var rawPoints: [Int] = []
             
             for item in parent.allItems {
                 switch item {
                 case .todo(let t):
                     if t.latitude.isNaN || t.longitude.isNaN { continue } // [NEW] NaN Guard
                     allItemsToProcess.append(item)
-                    rawPoints.append(Int32(t.int_lat))
-                    rawPoints.append(Int32(t.int_long))
+                    rawPoints.append(t.int_lat)
+                    rawPoints.append(t.int_long)
                 case .history(let log):
                     if log.latitude.isNaN || log.longitude.isNaN { continue } // [NEW] NaN Guard
                     allItemsToProcess.append(item)
-                    rawPoints.append(Int32(log.int_lat))
-                    rawPoints.append(Int32(log.int_long))
+                    rawPoints.append(log.int_lat)
+                    rawPoints.append(log.int_long)
                 case .userLocation(let coord):
                     if coord.latitude.isNaN || coord.longitude.isNaN { continue } // [NEW] NaN Guard
                     allItemsToProcess.append(item)
-                    rawPoints.append(Int32(coord.latitude * 100_000))
-                    rawPoints.append(Int32(coord.longitude * 100_000))
+                    rawPoints.append(Int(coord.latitude * 100_000))
+                    rawPoints.append(Int(coord.longitude * 100_000))
                 default: break
                 }
             }
@@ -345,8 +345,8 @@ struct KakaoMapView: UIViewRepresentable {
             if let target = parent.creatingTodoLocation {
                 let newItem = ToDoItem(todo_name: "New Entry", latitude: target.latitude, longitude: target.longitude)
                 allItemsToProcess.append(.todo(newItem))
-                rawPoints.append(Int32(target.latitude * 100_000))
-                rawPoints.append(Int32(target.longitude * 100_000))
+                rawPoints.append(Int(target.latitude * 100_000))
+                rawPoints.append(Int(target.longitude * 100_000))
             }
             
             // 3. Request WASM Clustering
@@ -414,7 +414,7 @@ struct KakaoMapView: UIViewRepresentable {
         }
 
         @MainActor
-        private func renderClusters(mapView: KakaoMap, clusterResult: [Int32], allItems: [UnifiedMapItem]) {
+        private func renderClusters(mapView: KakaoMap, clusterResult: [Int], allItems: [UnifiedMapItem]) {
             let labelManager = mapView.getLabelManager()
             let layer = labelManager.getLabelLayer(layerID: "todoLayer") ?? labelManager.addLabelLayer(option: LabelLayerOptions(layerID: "todoLayer", competitionType: .none, competitionUnit: .poi, orderType: .rank, zOrder: 20000))
             guard let validLayer = layer else { return }
